// src/types.d.ts
export type BandsType = {
  id: number
  title: string
  description: string
  videourl: string
  price: number | null
  coverImg: string
  category: string
  location: string
  locationImg: string
  openSpots: null | number
  likes: number
  websiteurl: string
  youtube: string
  facebook: string
  instagram: string
  twitter: string
  spotify: string
  deezer: string
  apple: string
  soundcloud: string
  tidal: string
  wiki: string
  ticketmaster: string
  tiktok: string
  napster: string
  bandera: string | undefined
}